requirements.txt
fastapi
uvicorn
python-dotenv
slowapi
pydantic
email-validator